Decompose cli build commands into more manageable files
